commit to merge with remote branch
